VehicleRoutingProblems

VehicleRoutingProblems.create_conclution_opt_solomon — Method

create_conclution_opt_solomon()

Create the csv file contain all solomon instances r101-rc208 Output name: data/optsolomon/`nametype(r1, r2, ...)/name-num_customer`

VehicleRoutingProblems.create_csv_min_all_distance — Method

create_csv_min_all_distance()

create csv file contains minimum distance of all runs using PSO

VehicleRoutingProblems.dataframe_group_opt_solomon — Method

dataframe_group_opt_solomon(df::DataFrames)

Find the group by the number of customers (25, 50, 100) of solomon instances

Return 3 groups of isntances gb[1], gb[2], gb[3] Each group will contain all instrnce of name_type

For example:

 if name_type is `r1` each group will have r101, r102, ..., r112

VehicleRoutingProblems.dataframe_opt_solomon — Method

dataframe_opt_solomon(name_type::String)

Find the group by the number of customers (25, 50, 100) of solomon instances

VehicleRoutingProblems.heuristic_single_min — Method

try to rearrenge sch by processing time. input:: if all_jobs is [] means condidates are 1, 2, ..., n Returns: array – sch that can not be processed in time.

VehicleRoutingProblems.point_distance — Method

point_distance(p1, p2)

p1: coordiante point in the form of [x1, y1]

p2: coordiante point in the form of [x2, y2]

Return euclidian distance from (x1, y1) -> (x2, y2)

VehicleRoutingProblems.read_route — Method

read_route(file_name::String)

return route representation from file text

For example:

Text file

1 2 4 5

8 9 3 6

Return Vector[0, 1, 2, 4, 5, 0, 8, 9, 3, 6, 0] where 0 represented depot

VehicleRoutingProblems.seperate_route — Method

seperate_route(solution::Solution)

return Vector of route with length = number of vehicles

Example

solution = read_solution("particle_swarm\total_distance\case1\r101\15\1\r101-1.txt", "r101")
seperate_route(solution)

VehicleRoutingProblems.total_route — Method

total_route(solution::Solution)

Return the number of route